+++ src/main/java/isthere/spring/dao/ShopDao.java
@@ -7,7 +7,7 @@ import isthere.spring.vo.Shop;
 public interface ShopDao {
     
     public String addShop(Shop shop);
-    public ArrayList<Shop> scanShop(int dist);
+    public ArrayList<Shop> scanShop(float dist);
      
 }
